# Improve `make_ffi_call` performance (part 2)