This further improves performance:
```
flipper/direct          time:   [48.326 ns 48.469 ns 48.638 ns]
                        thrpt:  [20.560 Melem/s 20.632 Melem/s 20.693 Melem/s]
flipper/transaction     time:   [57.682 ns 57.933 ns 58.219 ns]
                        thrpt:  [17.177 Melem/s 17.261 Melem/s 17.336 Melem/s]
example-wallet/execute-only
                        time:   [535.60 ns 537.22 ns 539.44 ns]
                        thrpt:  [1.8538 Melem/s 1.8614 Melem/s 1.8671 Melem/s]
```